requirements.txt
fastapi
uvicorn
python-dotenv
httpx
requests
yfinance
pandas
numpy
schwab-py
python-dateutil
aiofiles
